ci(all): upgrade to babel 7 and typescript
