+++ package.json
@@ -7,12 +7,16 @@
   "author": "Felix Wu",
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
+    "@babel/cli": "^7.6.2",
+    "@babel/core": "^7.6.2",
+    "@babel/preset-env": "^7.6.2",
+    "@babel/preset-typescript": "^7.6.0",
+    "@types/jest": "^24.0.18",
+    "@types/node": "^12.7.11",
+    "babel-jest": "^24.9.0",
     "babel-plugin-module-resolver": "^3.1.1",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-flow": "^6.23.0",
-    "flow-bin": "^0.77.0",
-    "jest": "^23.4.2"
+    "jest": "^24.9.0",
+    "ts-jest": "^24.1.0"
   },
   "keywords": [
     "markdown",
@@ -25,19 +29,23 @@
   ],
   "scripts": {
     "build": "babel src/ -d lib/",
-    "test": "jest",
-    "flow": "flow check"
+    "test": "jest"
   },
   "jest": {
     "verbose": true,
     "testEnvironment": "node",
+    "testMatch": [
+      "**/__tests__/**/*.[jt]s?(x)",
+      "**/?(*.)+(spec|test).[jt]s?(x)"
+    ],
     "transform": {
-      "^.+\\.js$": "babel-jest"
+      "^.+\\.ts$": "babel-jest"
     },
     "globals": {
       "NODE_ENV": "test"
     },
     "moduleFileExtensions": [
+      "ts",
       "js"
     ],
     "moduleDirectories": [
@@ -51,7 +59,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-push": "yarn flow && yarn test"
+      "pre-push": "yarn test"
     }
   }
 }
