+++ src/util/constants.ts
@@ -1,6 +1,5 @@
 /**
  * Constants
- * @flow
  */
 
 const HEADER_PREFIX: string = '#';
@@ -17,12 +16,12 @@ const SECTION_LINE_BREAK = '\n';
 const HORIZONTAL_RULE = '---';
 
 export {
-  HEADER_PREFIX,
-  EMPHASIS_ITALICS,
-  EMPHASIS_BOLD,
-  EMPHASIS_STRIKETHROUGH,
-  UNORDERED_LIST_PREFIX,
-  LINE_BREAK,
-  SECTION_LINE_BREAK,
-  HORIZONTAL_RULE
+	HEADER_PREFIX,
+	EMPHASIS_ITALICS,
+	EMPHASIS_BOLD,
+	EMPHASIS_STRIKETHROUGH,
+	UNORDERED_LIST_PREFIX,
+	LINE_BREAK,
+	SECTION_LINE_BREAK,
+	HORIZONTAL_RULE
 };
